Rename: uri를 File로 변환하는 공용 함수 위치 이동
